yubico.yubikey_usb_hid: add missing import
When a timeout occurs in the method YubiKeyUSBHID._waitfor, instead of
throwing the exception, yubikey.YubiKeyTimeout as expected, Python
throws a NameError instead because the yubikey module has not been
imported (only the YubiKey class).

The bug can be reproduced by altering the configure_nist_test_key;
setting require_button to True in the call to
Cfg.mode_challenge_response on line 20, running it and the actual test,
then letting the YubiKey time out instead of pressing the button.

The alternate fix to this would be to import the YubiKeyTimeout
exception class in addition to the YubiKey class.