yubico.yubikey: Fix typo in YubiKeyTimeout
When a yubikey.YubiKeyTimeout exception is created to be thrown, Python
throws a NameError instead because the constructor uses an undefined
variable..

The bug can be reproduced by altering the configure_nist_test_key;
setting require_button to True in the call to
Cfg.mode_challenge_response on line 20, running it and the actual test,
then letting the YubiKey time out instead of pressing the button.

The constructor has a parameter, 'value', which was probably intended to
be called 'reason' (or the constructor meant to pass 'value' instead).
This patch renames the constructor parameter to make it consistent with
other modules' usage of the class.